feat: switch prefer-repository-shorthand to repository-shorthand
