+++ README.md
@@ -113,17 +113,17 @@ The default settings don't conflict, and Prettier plugins can quickly fix up ord
 🔧 Automatically fixable by the [`--fix` CLI option](https://eslint.org/docs/user-guide/command-line-interface#--fix).\
 💡 Manually fixable by [editor suggestions](https://eslint.org/docs/latest/use/core-concepts#rule-suggestions).
 
-| Name                                                                     | Description                                                                                     | 💼 | 🔧 | 💡 |
-| :----------------------------------------------------------------------- | :---------------------------------------------------------------------------------------------- | :- | :- | :- |
-| [order-properties](docs/rules/order-properties.md)                       | Package properties must be declared in standard order                                           | ✅  | 🔧 |    |
-| [prefer-repository-shorthand](docs/rules/prefer-repository-shorthand.md) | Enforce shorthand declaration for GitHub repository.                                            | ✅  | 🔧 |    |
-| [sort-collections](docs/rules/sort-collections.md)                       | Dependencies, scripts, and configuration values must be declared in alphabetical order.         | ✅  | 🔧 |    |
-| [unique-dependencies](docs/rules/unique-dependencies.md)                 | Enforce that if repository directory is specified, it matches the path to the package.json file | ✅  |    | 💡 |
-| [valid-local-dependency](docs/rules/valid-local-dependency.md)           | Checks existence of local dependencies in the package.json                                      | ✅  |    |    |
-| [valid-name](docs/rules/valid-name.md)                                   | Enforce that package names are valid npm package names                                          | ✅  |    | 💡 |
-| [valid-package-def](docs/rules/valid-package-def.md)                     | Enforce that package.json has all properties required by the npm spec                           | ✅  |    |    |
-| [valid-repository-directory](docs/rules/valid-repository-directory.md)   | Enforce that if repository directory is specified, it matches the path to the package.json file | ✅  |    | 💡 |
-| [valid-version](docs/rules/valid-version.md)                             | Enforce that package versions are valid semver specifiers                                       | ✅  |    | 💡 |
+| Name                                                                   | Description                                                                                     | 💼 | 🔧 | 💡 |
+| :--------------------------------------------------------------------- | :---------------------------------------------------------------------------------------------- | :- | :- | :- |
+| [order-properties](docs/rules/order-properties.md)                     | Package properties must be declared in standard order                                           | ✅  | 🔧 |    |
+| [repository-shorthand](docs/rules/repository-shorthand.md)             | Enforce either object or shorthand declaration for repository.                                  | ✅  | 🔧 |    |
+| [sort-collections](docs/rules/sort-collections.md)                     | Dependencies, scripts, and configuration values must be declared in alphabetical order.         | ✅  | 🔧 |    |
+| [unique-dependencies](docs/rules/unique-dependencies.md)               | Enforce that if repository directory is specified, it matches the path to the package.json file | ✅  |    | 💡 |
+| [valid-local-dependency](docs/rules/valid-local-dependency.md)         | Checks existence of local dependencies in the package.json                                      | ✅  |    |    |
+| [valid-name](docs/rules/valid-name.md)                                 | Enforce that package names are valid npm package names                                          | ✅  |    | 💡 |
+| [valid-package-def](docs/rules/valid-package-def.md)                   | Enforce that package.json has all properties required by the npm spec                           | ✅  |    |    |
+| [valid-repository-directory](docs/rules/valid-repository-directory.md) | Enforce that if repository directory is specified, it matches the path to the package.json file | ✅  |    | 💡 |
+| [valid-version](docs/rules/valid-version.md)                           | Enforce that package versions are valid semver specifiers                                       | ✅  |    | 💡 |
 
 <!-- end auto-generated rules list -->
 <!-- prettier-ignore-end -->
