+++ src/plugin.ts
@@ -1,7 +1,7 @@
 import { createRequire } from "node:module";
 
 import { rule as orderProperties } from "./rules/order-properties.js";
-import { rule as preferRepositoryShorthand } from "./rules/prefer-repository-shorthand.js";
+import { rule as preferRepositoryShorthand } from "./rules/repository-shorthand.js";
 import { rule as sortCollections } from "./rules/sort-collections.js";
 import { rule as uniqueDependencies } from "./rules/unique-dependencies.js";
 import { rule as validLocalDependency } from "./rules/valid-local-dependency.js";
@@ -19,7 +19,7 @@ const { name, version } = require("../package.json") as {
 
 const rules = {
 	"order-properties": orderProperties,
-	"prefer-repository-shorthand": preferRepositoryShorthand,
+	"repository-shorthand": preferRepositoryShorthand,
 	"sort-collections": sortCollections,
 	"unique-dependencies": uniqueDependencies,
 	"valid-local-dependency": validLocalDependency,
