--- src/tests/rules/prefer-repository-shorthand.test.ts
@@ -1,132 +1,0 @@
-import { rule } from "../../rules/prefer-repository-shorthand.js";
-import { ruleTester } from "./ruleTester.js";
-
-ruleTester.run("prefer-repository-shorthand", rule, {
-	invalid: [
-		{
-			code: `{
-			"repository": {
-				"type": "git",
-				"url": "https://github.com/JoshuaKGoldberg/create-typescript-app"
-			}
-		}`,
-			errors: [
-				{
-					message:
-						"Prefer a shorthand locator for a GitHub repository.",
-					type: "JSONObjectExpression",
-				},
-			],
-			filename: "package.json",
-			output: `{
-			"repository": "JoshuaKGoldberg/create-typescript-app"
-		}`,
-		},
-		{
-			code: `{
-			"repository": {
-			  "type": "git",
-			  "url": "git+https://github.com/JoshuaKGoldberg/create-typescript-app.git"
-			}
-		}`,
-			errors: [
-				{
-					message:
-						"Prefer a shorthand locator for a GitHub repository.",
-					type: "JSONObjectExpression",
-				},
-			],
-			filename: "package.json",
-			output: `{
-			"repository": "JoshuaKGoldberg/create-typescript-app"
-		}`,
-		},
-		{
-			code: `{
-			"repository": {
-				"type": "git",
-				"url": "git+ssh://git@github.com/JoshuaKGoldberg/create-typescript-app.git"
-			}
-		}`,
-			errors: [
-				{
-					message:
-						"Prefer a shorthand locator for a GitHub repository.",
-					type: "JSONObjectExpression",
-				},
-			],
-			filename: "package.json",
-			output: `{
-			"repository": "JoshuaKGoldberg/create-typescript-app"
-		}`,
-		},
-		{
-			code: `{
-			"repository": "https://github.com/eslint/eslint"
-		}`,
-			errors: [
-				{
-					message:
-						"Prefer a shorthand locator for a GitHub repository.",
-					type: "JSONLiteral",
-				},
-			],
-			filename: "package.json",
-			output: `{
-			"repository": "eslint/eslint"
-		}`,
-		},
-	],
-	valid: [
-		{
-			code: `{
-			"repository": {
-				"type": "git",
-				"url": "https://github.com/facebook/react.git",
-				"directory": "packages/react"
-			}
-		}`,
-			filename: "package.json",
-		},
-		{
-			code: `{
-			"repository": {
-				"url": "https://github.com/facebook/react.git",
-			}
-		}`,
-			filename: "package.json",
-		},
-		{
-			code: `{
-			"repository": "browserslist/browserslist"
-		}`,
-			filename: "package.json",
-		},
-		{
-			code: `{
-			"repository": {
-				"type": "git",
-				"url": "https://gitlab.com/gitlab/gitlab"
-			}
-		}`,
-			filename: "package.json",
-		},
-		{
-			code: `{
-			"repository": {
-				"type": "git",
-				"url": "https://gitlab.com/a/b/c"
-			}
-		}`,
-			filename: "package.json",
-		},
-		{
-			code: `{
-		        "outer": {
-		            "repository": "https://github.com/a/b"
-		        }
-		    }`,
-			filename: "package.json",
-		},
-	],
-});
